refactor(phase4): split compiler monoliths into focused submodules
Extract service-level token generators out of codegen.rs (8.5 kloc →
7.2 kloc) into codegen/service_tokens.rs (1.26 kloc), and extract all
serde document types out of eon_parser.rs (11.3 kloc → 9.5 kloc) into
eon_parser/documents.rs (1.74 kloc).

Each new submodule carries its own minimal import set. The parent
modules re-import via `use self::submod::*` so call-sites are
unaffected. `LoadedService` remains in eon_parser.rs (it is the
parser's public return type, not a document type).
