+++ crates/rest_macro_core/src/compiler/eon_parser.rs
@@ -1,4 +1,4 @@
-use std::{
+﻿use std::{
     collections::{BTreeMap, HashMap, HashSet},
     env, fs,
     path::{Component, Path, PathBuf},
@@ -9,7 +9,6 @@ use heck::ToSnakeCase;
 use proc_macro2::Span;
 use quote::ToTokens;
 use rust_decimal::Decimal;
-use serde::de::{self, Deserializer, Error as _, MapAccess, Visitor};
 use serde_json::Value as JsonValue;
 use syn::{LitStr, Type};
 use uuid::Uuid;
@@ -75,1732 +74,20 @@ use crate::{
     },
 };
 
-pub struct LoadedService {
-    pub service: ServiceSpec,
-    pub include_path: String,
-}
-
-#[derive(serde::Deserialize)]
-struct ServiceDocument {
-    #[serde(default)]
-    module: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_enum_documents")]
-    enums: Vec<EnumDocument>,
-    #[serde(default, deserialize_with = "deserialize_mixin_documents")]
-    mixins: Vec<MixinDocument>,
-    #[serde(default)]
-    db: DbBackend,
-    #[serde(default)]
-    database: Option<DatabaseDocument>,
-    #[serde(default)]
-    build: Option<BuildDocument>,
-    #[serde(default)]
-    clients: Option<ClientsDocument>,
-    #[serde(default)]
-    logging: Option<LoggingDocument>,
-    #[serde(default)]
-    runtime: Option<RuntimeDocument>,
-    #[serde(default)]
-    storage: Option<StorageDocument>,
-    #[serde(default)]
-    authorization: Option<AuthorizationDocument>,
-    #[serde(default)]
-    tls: Option<TlsDocument>,
-    #[serde(default, rename = "static")]
-    static_config: Option<StaticConfigDocument>,
-    #[serde(default)]
-    security: SecurityDocument,
-    #[serde(deserialize_with = "deserialize_resource_documents")]
-    resources: Vec<ResourceDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct DatabaseDocument {
-    #[serde(default)]
-    engine: Option<DatabaseEngineDocument>,
-    #[serde(default)]
-    resilience: Option<DatabaseResilienceDocument>,
-}
-
-#[derive(serde::Deserialize)]
-struct DatabaseEngineDocument {
-    kind: String,
-    #[serde(default)]
-    path: Option<String>,
-    #[serde(default)]
-    encryption_key_env: Option<String>,
-    #[serde(default)]
-    encryption_key: Option<SecretRefDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct DatabaseResilienceDocument {
-    #[serde(default)]
-    profile: Option<String>,
-    #[serde(default)]
-    backup: Option<DatabaseBackupDocument>,
-    #[serde(default)]
-    replication: Option<DatabaseReplicationDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct DatabaseBackupDocument {
-    #[serde(default)]
-    required: Option<bool>,
-    #[serde(default)]
-    mode: Option<String>,
-    #[serde(default)]
-    target: Option<String>,
-    #[serde(default)]
-    verify_restore: Option<bool>,
-    #[serde(default)]
-    max_age: Option<String>,
-    #[serde(default)]
-    encryption_key_env: Option<String>,
-    #[serde(default)]
-    encryption_key: Option<SecretRefDocument>,
-    #[serde(default)]
-    retention: Option<DatabaseBackupRetentionDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct DatabaseBackupRetentionDocument {
-    #[serde(default)]
-    daily: Option<u32>,
-    #[serde(default)]
-    weekly: Option<u32>,
-    #[serde(default)]
-    monthly: Option<u32>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct DatabaseReplicationDocument {
-    #[serde(default)]
-    mode: Option<String>,
-    #[serde(default)]
-    read_routing: Option<String>,
-    #[serde(default)]
-    read_url_env: Option<String>,
-    #[serde(default)]
-    read_url: Option<SecretRefDocument>,
-    #[serde(default)]
-    max_lag: Option<String>,
-    #[serde(default)]
-    replicas_expected: Option<u32>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct SecurityDocument {
-    #[serde(default)]
-    requests: Option<RequestSecurityDocument>,
-    #[serde(default)]
-    cors: Option<CorsSecurityDocument>,
-    #[serde(default)]
-    trusted_proxies: Option<TrustedProxiesDocument>,
-    #[serde(default)]
-    rate_limits: Option<RateLimitsDocument>,
-    #[serde(default)]
-    access: Option<SecurityAccessDocument>,
-    #[serde(default)]
-    headers: Option<HeaderSecurityDocument>,
-    #[serde(default)]
-    auth: Option<AuthSecurityDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct SecurityAccessDocument {
-    #[serde(default)]
-    default_read: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct RequestSecurityDocument {
-    #[serde(default)]
-    json_max_bytes: Option<usize>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct CorsSecurityDocument {
-    #[serde(default)]
-    origins: Vec<String>,
-    #[serde(default)]
-    origins_env: Option<String>,
-    #[serde(default)]
-    allow_credentials: Option<bool>,
-    #[serde(default)]
-    allow_methods: Vec<String>,
-    #[serde(default)]
-    allow_headers: Vec<String>,
-    #[serde(default)]
-    expose_headers: Vec<String>,
-    #[serde(default)]
-    max_age_seconds: Option<usize>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct TrustedProxiesDocument {
-    #[serde(default)]
-    proxies: Vec<String>,
-    #[serde(default)]
-    proxies_env: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct RateLimitsDocument {
-    #[serde(default)]
-    login: Option<RateLimitRuleDocument>,
-    #[serde(default)]
-    register: Option<RateLimitRuleDocument>,
-}
-
-#[derive(serde::Deserialize)]
-struct RateLimitRuleDocument {
-    requests: u32,
-    window_seconds: u64,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct HeaderSecurityDocument {
-    #[serde(default)]
-    frame_options: Option<String>,
-    #[serde(default)]
-    content_type_options: Option<bool>,
-    #[serde(default)]
-    referrer_policy: Option<String>,
-    #[serde(default)]
-    hsts: Option<HstsDocument>,
-}
-
-#[derive(serde::Deserialize)]
-struct HstsDocument {
-    max_age_seconds: u64,
-    #[serde(default)]
-    include_subdomains: bool,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthSecurityDocument {
-    #[serde(default)]
-    issuer: Option<String>,
-    #[serde(default)]
-    audience: Option<String>,
-    #[serde(default)]
-    access_token_ttl_seconds: Option<i64>,
-    #[serde(default)]
-    require_email_verification: Option<bool>,
-    #[serde(default)]
-    verification_token_ttl_seconds: Option<i64>,
-    #[serde(default)]
-    password_reset_token_ttl_seconds: Option<i64>,
-    #[serde(default)]
-    jwt: Option<AuthJwtDocument>,
-    #[serde(default)]
-    jwt_secret: Option<SecretRefDocument>,
-    #[serde(default)]
-    claims: BTreeMap<String, AuthClaimMapValueDocument>,
-    #[serde(default)]
-    session_cookie: Option<SessionCookieDocument>,
-    #[serde(default)]
-    email: Option<AuthEmailDocument>,
-    #[serde(default)]
-    portal: Option<AuthUiPageDocument>,
-    #[serde(default)]
-    admin_dashboard: Option<AuthUiPageDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthJwtDocument {
-    #[serde(default)]
-    algorithm: Option<String>,
-    #[serde(default)]
-    active_kid: Option<String>,
-    #[serde(default)]
-    signing_key: Option<SecretRefDocument>,
-    #[serde(default)]
-    verification_keys: Vec<AuthJwtVerificationKeyDocument>,
-}
-
-#[derive(serde::Deserialize)]
-struct AuthJwtVerificationKeyDocument {
-    kid: String,
-    key: SecretRefDocument,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct SessionCookieDocument {
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    csrf_cookie_name: Option<String>,
-    #[serde(default)]
-    csrf_header_name: Option<String>,
-    #[serde(default)]
-    path: Option<String>,
-    #[serde(default)]
-    secure: Option<bool>,
-    #[serde(default)]
-    same_site: Option<String>,
-}
-
-#[derive(serde::Deserialize)]
-struct AuthEmailDocument {
-    from_email: String,
-    #[serde(default)]
-    from_name: Option<String>,
-    #[serde(default)]
-    reply_to: Option<String>,
-    #[serde(default)]
-    public_base_url: Option<String>,
-    provider: AuthEmailProviderDocument,
-}
-
-#[derive(serde::Deserialize)]
-struct AuthEmailProviderDocument {
-    kind: String,
-    #[serde(default)]
-    api_key_env: Option<String>,
-    #[serde(default)]
-    api_key: Option<SecretRefDocument>,
-    #[serde(default)]
-    api_base_url: Option<String>,
-    #[serde(default)]
-    connection_url_env: Option<String>,
-    #[serde(default)]
-    connection_url: Option<SecretRefDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct SecretRefDocument {
-    #[serde(default)]
-    env: Option<String>,
-    #[serde(default)]
-    env_or_file: Option<String>,
-    #[serde(default)]
-    systemd_credential: Option<String>,
-    #[serde(default)]
-    external: Option<ExternalSecretRefDocument>,
-}
-
-#[derive(serde::Deserialize)]
-struct ExternalSecretRefDocument {
-    provider: String,
-    locator: String,
-}
-
-#[derive(serde::Deserialize)]
-struct AuthUiPageDocument {
-    path: String,
-    #[serde(default)]
-    title: Option<String>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum AuthClaimMapValueDocument {
-    Type(AuthClaimTypeDocument),
-    Column(String),
-    Config(AuthClaimConfigDocument),
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthClaimConfigDocument {
-    #[serde(default)]
-    column: Option<String>,
-    #[serde(default, rename = "type")]
-    ty: Option<AuthClaimTypeDocument>,
-}
-
-#[derive(Clone, Copy, serde::Deserialize)]
-enum AuthClaimTypeDocument {
-    String,
-    I64,
-    Bool,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct LoggingDocument {
-    #[serde(default)]
-    filter_env: Option<String>,
-    #[serde(default)]
-    default_filter: Option<String>,
-    #[serde(default)]
-    timestamp: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct BuildDocument {
-    #[serde(default)]
-    target_cpu_native: Option<bool>,
-    #[serde(default)]
-    release: Option<ReleaseBuildDocument>,
-    #[serde(default)]
-    artifacts: Option<BuildArtifactsDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct ReleaseBuildDocument {
-    #[serde(default)]
-    lto: Option<BuildLtoDocument>,
-    #[serde(default)]
-    codegen_units: Option<u32>,
-    #[serde(default)]
-    strip_debug_symbols: Option<bool>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct BuildArtifactsDocument {
-    #[serde(default)]
-    binary: Option<BuildArtifactPathDocument>,
-    #[serde(default)]
-    bundle: Option<BuildArtifactPathDocument>,
-    #[serde(default)]
-    cache: Option<BuildCacheArtifactDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct BuildArtifactPathDocument {
-    #[serde(default)]
-    path: Option<String>,
-    #[serde(default)]
-    env: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct ClientsDocument {
-    #[serde(default)]
-    ts: Option<TsClientDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct TsClientDocument {
-    #[serde(default)]
-    output_dir: Option<BuildArtifactPathDocument>,
-    #[serde(default)]
-    package_name: Option<ClientValueDocument>,
-    #[serde(default)]
-    server_url: Option<String>,
-    #[serde(default)]
-    emit_js: Option<bool>,
-    #[serde(default)]
-    include_builtin_auth: Option<bool>,
-    #[serde(default)]
-    exclude_tables: Option<Vec<String>>,
-    #[serde(default)]
-    automation: Option<TsClientAutomationDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct TsClientAutomationDocument {
-    #[serde(default)]
-    on_build: Option<bool>,
-    #[serde(default)]
-    self_test: Option<bool>,
-    #[serde(default)]
-    self_test_report: Option<BuildArtifactPathDocument>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum ClientValueDocument {
-    Value(String),
-    Config(ClientValueConfigDocument),
-}
-
-#[derive(Default, serde::Deserialize)]
-struct ClientValueConfigDocument {
-    #[serde(default)]
-    value: Option<String>,
-    #[serde(default)]
-    env: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct BuildCacheArtifactDocument {
-    #[serde(default)]
-    root: Option<String>,
-    #[serde(default)]
-    env: Option<String>,
-    #[serde(default)]
-    cleanup: Option<String>,
-}
-
-#[derive(Clone, Copy, serde::Deserialize)]
-#[serde(untagged)]
-enum BuildLtoDocument {
-    Bool(bool),
-    Mode(BuildLtoModeDocument),
-}
-
-#[derive(Clone, Copy, serde::Deserialize)]
-enum BuildLtoModeDocument {
-    Thin,
-    Fat,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct RuntimeDocument {
-    #[serde(default)]
-    compression: Option<CompressionDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthorizationDocument {
-    #[serde(default)]
-    scopes: BTreeMap<String, AuthorizationScopeDocument>,
-    #[serde(default)]
-    permissions: BTreeMap<String, AuthorizationPermissionDocument>,
-    #[serde(default)]
-    templates: BTreeMap<String, AuthorizationTemplateDocument>,
-    #[serde(default)]
-    hybrid_enforcement: Option<AuthorizationHybridEnforcementDocument>,
-    #[serde(default)]
-    management_api: Option<AuthorizationManagementApiDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthorizationHybridEnforcementDocument {
-    #[serde(default)]
-    resources: BTreeMap<String, AuthorizationHybridResourceDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthorizationHybridResourceDocument {
-    #[serde(default)]
-    scope: Option<String>,
-    #[serde(default)]
-    scope_field: Option<String>,
-    #[serde(default)]
-    scope_sources: Option<AuthorizationHybridScopeSourcesDocument>,
-    #[serde(default)]
-    actions: Vec<AuthorizationActionDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthorizationHybridScopeSourcesDocument {
-    #[serde(default)]
-    item: Option<bool>,
-    #[serde(default)]
-    collection_filter: Option<bool>,
-    #[serde(default)]
-    nested_parent: Option<bool>,
-    #[serde(default)]
-    create_payload: Option<bool>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthorizationManagementApiDocument {
-    #[serde(default)]
-    enabled: Option<bool>,
-    #[serde(default)]
-    mount: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthorizationScopeDocument {
-    #[serde(default)]
-    description: Option<String>,
-    #[serde(default)]
-    parent: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthorizationPermissionDocument {
-    #[serde(default)]
-    description: Option<String>,
-    #[serde(default)]
-    actions: Vec<AuthorizationActionDocument>,
-    #[serde(default)]
-    resources: Vec<String>,
-    #[serde(default)]
-    scopes: Vec<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct AuthorizationTemplateDocument {
-    #[serde(default)]
-    description: Option<String>,
-    #[serde(default)]
-    permissions: Vec<String>,
-    #[serde(default)]
-    scopes: Vec<String>,
-}
-
-#[derive(Clone, Copy, serde::Deserialize)]
-enum AuthorizationActionDocument {
-    Read,
-    Create,
-    Update,
-    Delete,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct CompressionDocument {
-    #[serde(default)]
-    enabled: Option<bool>,
-    #[serde(default)]
-    static_precompressed: Option<bool>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct StorageDocument {
-    #[serde(default)]
-    backends: Vec<StorageBackendDocument>,
-    #[serde(default)]
-    public_mounts: Vec<StoragePublicMountDocument>,
-    #[serde(default)]
-    uploads: Vec<StorageUploadDocument>,
-    #[serde(default)]
-    s3_compat: Option<StorageS3CompatDocument>,
-}
-
-#[derive(serde::Deserialize)]
-struct StorageBackendDocument {
-    name: String,
-    kind: String,
-    dir: String,
-}
-
-#[derive(serde::Deserialize)]
-struct StoragePublicMountDocument {
-    mount: String,
-    backend: String,
-    #[serde(default)]
-    prefix: Option<String>,
-    #[serde(default)]
-    cache: Option<String>,
-}
-
-#[derive(serde::Deserialize)]
-struct StorageUploadDocument {
-    name: String,
-    path: String,
-    backend: String,
-    #[serde(default)]
-    prefix: Option<String>,
-    #[serde(default)]
-    max_bytes: Option<usize>,
-    #[serde(default)]
-    require_auth: Option<bool>,
-    #[serde(default)]
-    roles: Vec<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct StorageS3CompatDocument {
-    #[serde(default)]
-    mount: Option<String>,
-    #[serde(default)]
-    buckets: Vec<StorageS3CompatBucketDocument>,
-}
-
-#[derive(serde::Deserialize)]
-struct StorageS3CompatBucketDocument {
-    name: String,
-    backend: String,
-    #[serde(default)]
-    prefix: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct TlsDocument {
-    #[serde(default)]
-    cert_path: Option<String>,
-    #[serde(default)]
-    key_path: Option<String>,
-    #[serde(default)]
-    cert_path_env: Option<String>,
-    #[serde(default)]
-    key_path_env: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct StaticConfigDocument {
-    #[serde(default)]
-    mounts: Vec<StaticMountDocument>,
-}
-
-#[derive(serde::Deserialize)]
-struct StaticMountDocument {
-    mount: String,
-    dir: String,
-    #[serde(default)]
-    mode: Option<String>,
-    #[serde(default)]
-    index_file: Option<String>,
-    #[serde(default)]
-    fallback_file: Option<String>,
-    #[serde(default)]
-    cache: Option<String>,
-}
-
-#[derive(serde::Deserialize)]
-struct ResourceDocument {
-    name: String,
-    #[serde(default)]
-    table: Option<String>,
-    #[serde(default)]
-    api_name: Option<String>,
-    #[serde(default)]
-    id_field: Option<String>,
-    #[serde(default)]
-    access: ResourceAccessDocument,
-    #[serde(default)]
-    roles: RoleRequirements,
-    #[serde(default)]
-    policies: RowPoliciesDocument,
-    #[serde(default)]
-    list: ListConfigDocument,
-    #[serde(default)]
-    api: Option<ResourceApiDocument>,
-    #[serde(default, rename = "use")]
-    use_mixins: Vec<String>,
-    #[serde(default)]
-    indexes: Vec<IndexDocument>,
-    #[serde(default)]
-    many_to_many: Vec<ManyToManyDocument>,
-    #[serde(default)]
-    actions: Vec<ResourceActionDocument>,
-    #[serde(deserialize_with = "deserialize_field_documents")]
-    fields: Vec<FieldDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct ResourceAccessDocument {
-    #[serde(default)]
-    read: Option<String>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-struct MixinDocument {
-    name: String,
-    #[serde(default)]
-    indexes: Vec<IndexDocument>,
-    #[serde(default, deserialize_with = "deserialize_field_documents")]
-    fields: Vec<FieldDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct ResourceApiDocument {
-    #[serde(
-        default,
-        deserialize_with = "deserialize_api_field_projection_documents"
-    )]
-    fields: Vec<ApiFieldProjectionDocument>,
-    #[serde(default)]
-    default_context: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_response_context_documents")]
-    contexts: Vec<ResponseContextDocument>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ResourceActionDocument {
-    name: String,
-    #[serde(default)]
-    path: Option<String>,
-    #[serde(default)]
-    target: Option<String>,
-    #[serde(default)]
-    method: Option<String>,
-    behavior: ResourceActionBehaviorDocument,
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ResourceActionBehaviorDocument {
-    kind: String,
-    #[serde(default)]
-    set: BTreeMap<String, ResourceActionAssignmentValueDocument>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(untagged)]
-enum ResourceActionAssignmentValueDocument {
-    Input(ResourceActionInputValueDocument),
-    Literal(JsonValue),
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ResourceActionInputValueDocument {
-    input: String,
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ManyToManyDocument {
-    name: String,
-    target: String,
-    through: String,
-    source_field: String,
-    target_field: String,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct ListConfigDocument {
-    #[serde(default)]
-    default_limit: Option<u32>,
-    #[serde(default)]
-    max_limit: Option<u32>,
-}
-
-#[derive(Default, serde::Deserialize)]
-struct RowPoliciesDocument {
-    #[serde(default = "default_admin_bypass")]
-    admin_bypass: bool,
-    #[serde(default)]
-    read: Option<FilterPoliciesDocument>,
-    #[serde(default, deserialize_with = "deserialize_create_policies_document")]
-    create: Option<CreatePoliciesDocument>,
-    #[serde(default)]
-    update: Option<FilterPoliciesDocument>,
-    #[serde(default)]
-    delete: Option<FilterPoliciesDocument>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum CreatePoliciesDocument {
-    Structured(CreatePoliciesGroupDocument),
-    Assignments(ScopePoliciesDocument),
-}
-
-#[derive(Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct CreatePoliciesGroupDocument {
-    #[serde(default)]
-    assign: Option<ScopePoliciesDocument>,
-    #[serde(default)]
-    require: Option<FilterPoliciesDocument>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum FilterPoliciesDocument {
-    Group(FilterPolicyGroupDocument),
-    Many(Vec<FilterPoliciesDocument>),
-    Single(PolicyEntryDocument),
-}
-
-#[derive(Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct FilterPolicyGroupDocument {
-    #[serde(default)]
-    all_of: Option<Vec<FilterPoliciesDocument>>,
-    #[serde(default)]
-    any_of: Option<Vec<FilterPoliciesDocument>>,
-    #[serde(default)]
-    not: Option<Box<FilterPoliciesDocument>>,
-    #[serde(default)]
-    exists: Option<ExistsPolicyDocument>,
-}
 
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum ScopePoliciesDocument {
-    Many(Vec<PolicyEntryDocument>),
-    Single(PolicyEntryDocument),
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum PolicyEntryDocument {
-    Shorthand(String),
-    Rule(PolicyRuleDocument),
-    Legacy(LegacyRowPolicyDocument),
-}
-
-#[derive(Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ExistsPolicyDocument {
-    resource: String,
-    #[serde(default, rename = "where")]
-    condition: Option<ExistsPolicyEntriesDocument>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum ExistsPolicyEntriesDocument {
-    Group(ExistsPolicyGroupDocument),
-    Many(Vec<ExistsPolicyEntriesDocument>),
-    Single(ExistsPolicyEntryDocument),
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum ExistsPolicyEntryDocument {
-    Shorthand(String),
-    Rule(ExistsPolicyRuleDocument),
-    Legacy(LegacyRowPolicyDocument),
-}
-
-#[derive(Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ExistsPolicyGroupDocument {
-    #[serde(default)]
-    all_of: Option<Vec<ExistsPolicyEntriesDocument>>,
-    #[serde(default)]
-    any_of: Option<Vec<ExistsPolicyEntriesDocument>>,
-    #[serde(default)]
-    not: Option<Box<ExistsPolicyEntriesDocument>>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct LegacyRowPolicyDocument {
-    kind: String,
-    field: String,
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(untagged)]
-enum PolicyComparisonValueDocument {
-    String(String),
-    Integer(i64),
-    Bool(bool),
-}
-
-#[derive(serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct PolicyRuleDocument {
-    field: String,
-    #[serde(default)]
-    equals: Option<PolicyComparisonValueDocument>,
-    #[serde(default)]
-    is_null: bool,
-    #[serde(default)]
-    is_not_null: bool,
-    #[serde(default)]
-    value: Option<String>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ExistsPolicyRuleDocument {
-    field: String,
-    #[serde(default)]
-    equals: Option<PolicyComparisonValueDocument>,
-    #[serde(default)]
-    is_null: bool,
-    #[serde(default)]
-    is_not_null: bool,
-    #[serde(default)]
-    equals_field: Option<String>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-struct FieldDocument {
-    name: String,
-    #[serde(default)]
-    api_name: Option<String>,
-    #[serde(rename = "type")]
-    ty: FieldTypeDocument,
-    #[serde(default)]
-    items: Option<FieldTypeDocument>,
-    #[serde(default, deserialize_with = "deserialize_field_documents")]
-    fields: Vec<FieldDocument>,
-    #[serde(default)]
-    nullable: bool,
-    #[serde(default)]
-    id: bool,
-    #[serde(default)]
-    generated: GeneratedValue,
-    #[serde(default)]
-    unique: bool,
-    #[serde(default)]
-    transforms: Vec<String>,
-    #[serde(default)]
-    relation: Option<RelationDocument>,
-    #[serde(default, rename = "garde")]
-    validate: Option<FieldValidationDocument>,
-    #[serde(default, rename = "validate")]
-    legacy_validate: Option<JsonValue>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct IndexDocument {
-    fields: Vec<String>,
-    #[serde(default)]
-    unique: bool,
-}
-
-#[derive(Clone, serde::Deserialize)]
-struct EnumDocument {
-    name: String,
-    values: Vec<String>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-struct RelationDocument {
-    references: String,
-    #[serde(default)]
-    on_delete: Option<String>,
-    #[serde(default)]
-    nested_route: bool,
-}
-
-#[derive(Clone, Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct FieldValidationDocument {
-    #[serde(default)]
-    ascii: bool,
-    #[serde(default)]
-    alphanumeric: bool,
-    #[serde(default)]
-    email: bool,
-    #[serde(default)]
-    url: bool,
-    #[serde(default)]
-    ip: bool,
-    #[serde(default)]
-    ipv4: bool,
-    #[serde(default)]
-    ipv6: bool,
-    #[serde(default)]
-    phone_number: bool,
-    #[serde(default)]
-    credit_card: bool,
-    #[serde(default)]
-    required: bool,
-    #[serde(default)]
-    dive: bool,
-    #[serde(default)]
-    contains: Option<String>,
-    #[serde(default)]
-    prefix: Option<String>,
-    #[serde(default)]
-    suffix: Option<String>,
-    #[serde(default)]
-    pattern: Option<String>,
-    #[serde(default)]
-    length: Option<LengthValidationDocument>,
-    #[serde(default)]
-    range: Option<RangeValidationDocument>,
-    #[serde(default)]
-    inner: Option<Box<FieldValidationDocument>>,
-}
-
-#[derive(Clone, Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct LengthValidationDocument {
-    #[serde(default)]
-    min: Option<usize>,
-    #[serde(default)]
-    max: Option<usize>,
-    #[serde(default)]
-    equal: Option<usize>,
-    #[serde(default)]
-    mode: Option<String>,
-}
-
-#[derive(Clone, Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct RangeValidationDocument {
-    #[serde(default)]
-    min: Option<NumericBoundDocument>,
-    #[serde(default)]
-    max: Option<NumericBoundDocument>,
-    #[serde(default)]
-    equal: Option<NumericBoundDocument>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(untagged)]
-enum NumericBoundDocument {
-    Integer(i64),
-    Float(f64),
-}
-
-#[derive(Clone, serde::Deserialize)]
-#[serde(untagged)]
-enum FieldTypeDocument {
-    Scalar(ScalarType),
-    Rust(String),
-}
-
-#[derive(Clone, Copy, serde::Deserialize)]
-enum ScalarType {
-    String,
-    I32,
-    I64,
-    F32,
-    F64,
-    Bool,
-    DateTime,
-    Date,
-    Time,
-    Uuid,
-    Decimal,
-    Json,
-    JsonObject,
-    JsonArray,
-    List,
-    Object,
-}
-
-#[derive(serde::Deserialize)]
-struct ResourceMapValueDocument {
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    table: Option<String>,
-    #[serde(default)]
-    api_name: Option<String>,
-    #[serde(default)]
-    id_field: Option<String>,
-    #[serde(default)]
-    access: ResourceAccessDocument,
-    #[serde(default)]
-    roles: RoleRequirements,
-    #[serde(default)]
-    policies: RowPoliciesDocument,
-    #[serde(default)]
-    list: ListConfigDocument,
-    #[serde(default)]
-    api: Option<ResourceApiDocument>,
-    #[serde(default, rename = "use")]
-    use_mixins: Vec<String>,
-    #[serde(default)]
-    indexes: Vec<IndexDocument>,
-    #[serde(default)]
-    many_to_many: Vec<ManyToManyDocument>,
-    #[serde(default)]
-    actions: Vec<ResourceActionDocument>,
-    #[serde(deserialize_with = "deserialize_field_documents")]
-    fields: Vec<FieldDocument>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum MixinMapValueDocument {
-    Fields(Vec<FieldDocument>),
-    Config(MixinConfigDocument),
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-#[allow(clippy::large_enum_variant)]
-enum FieldMapValueDocument {
-    Type(FieldTypeDocument),
-    Config(FieldMapConfigDocument),
-}
-
-#[derive(serde::Deserialize)]
-struct FieldMapConfigDocument {
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    api_name: Option<String>,
-    #[serde(rename = "type")]
-    ty: FieldTypeDocument,
-    #[serde(default)]
-    items: Option<FieldTypeDocument>,
-    #[serde(default, deserialize_with = "deserialize_field_documents")]
-    fields: Vec<FieldDocument>,
-    #[serde(default)]
-    nullable: bool,
-    #[serde(default)]
-    id: bool,
-    #[serde(default)]
-    generated: GeneratedValue,
-    #[serde(default)]
-    unique: bool,
-    #[serde(default)]
-    transforms: Vec<String>,
-    #[serde(default)]
-    relation: Option<RelationDocument>,
-    #[serde(default, rename = "garde")]
-    validate: Option<FieldValidationDocument>,
-    #[serde(default, rename = "validate")]
-    legacy_validate: Option<JsonValue>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-struct ApiFieldProjectionDocument {
-    name: String,
-    #[serde(default)]
-    from: Option<String>,
-    #[serde(default)]
-    template: Option<String>,
-}
-
-#[derive(Clone, serde::Deserialize)]
-struct ResponseContextDocument {
-    name: String,
-    #[serde(default)]
-    fields: Vec<String>,
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum EnumMapValueDocument {
-    Values(Vec<String>),
-    Config(EnumConfigDocument),
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum ApiFieldProjectionMapValueDocument {
-    From(String),
-    Config(ApiFieldProjectionConfigDocument),
-}
-
-#[derive(serde::Deserialize)]
-#[serde(untagged)]
-enum ResponseContextMapValueDocument {
-    Fields(Vec<String>),
-    Config(ResponseContextConfigDocument),
-}
-
-#[derive(serde::Deserialize)]
-struct ApiFieldProjectionConfigDocument {
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    from: Option<String>,
-    #[serde(default)]
-    template: Option<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct MixinConfigDocument {
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    indexes: Vec<IndexDocument>,
-    #[serde(default, deserialize_with = "deserialize_field_documents")]
-    fields: Vec<FieldDocument>,
-}
-
-#[derive(Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct ResponseContextConfigDocument {
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    fields: Vec<String>,
-}
-
-#[derive(Default, serde::Deserialize)]
-#[serde(deny_unknown_fields)]
-struct EnumConfigDocument {
-    #[serde(default)]
-    name: Option<String>,
-    values: Vec<String>,
-}
-
-fn deserialize_enum_documents<'de, D>(deserializer: D) -> Result<Vec<EnumDocument>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    deserializer.deserialize_any(EnumDocumentsVisitor)
-}
-
-fn deserialize_mixin_documents<'de, D>(deserializer: D) -> Result<Vec<MixinDocument>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    deserializer.deserialize_any(MixinDocumentsVisitor)
-}
-
-fn deserialize_resource_documents<'de, D>(
-    deserializer: D,
-) -> Result<Vec<ResourceDocument>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    deserializer.deserialize_any(ResourceDocumentsVisitor)
-}
-
-fn deserialize_field_documents<'de, D>(deserializer: D) -> Result<Vec<FieldDocument>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    deserializer.deserialize_any(FieldDocumentsVisitor)
-}
-
-fn deserialize_api_field_projection_documents<'de, D>(
-    deserializer: D,
-) -> Result<Vec<ApiFieldProjectionDocument>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    deserializer.deserialize_any(ApiFieldProjectionDocumentsVisitor)
-}
-
-fn deserialize_response_context_documents<'de, D>(
-    deserializer: D,
-) -> Result<Vec<ResponseContextDocument>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    deserializer.deserialize_any(ResponseContextDocumentsVisitor)
-}
-
-impl ResourceMapValueDocument {
-    fn into_document<E>(self, key: String) -> Result<ResourceDocument, E>
-    where
-        E: de::Error,
-    {
-        if let Some(name) = self.name.as_deref()
-            && name != key
-        {
-            return Err(E::custom(format!(
-                "resource map entry `{key}` has mismatched `name` value `{name}`"
-            )));
-        }
-
-        Ok(ResourceDocument {
-            name: key,
-            table: self.table,
-            api_name: self.api_name,
-            id_field: self.id_field,
-            access: self.access,
-            roles: self.roles,
-            policies: self.policies,
-            list: self.list,
-            api: self.api,
-            use_mixins: self.use_mixins,
-            indexes: self.indexes,
-            many_to_many: self.many_to_many,
-            actions: self.actions,
-            fields: self.fields,
-        })
-    }
-}
-
-impl MixinMapValueDocument {
-    fn into_document<E>(self, key: String) -> Result<MixinDocument, E>
-    where
-        E: de::Error,
-    {
-        match self {
-            Self::Fields(fields) => Ok(MixinDocument {
-                name: key,
-                indexes: Vec::new(),
-                fields,
-            }),
-            Self::Config(mixin) => mixin.into_document::<E>(key),
-        }
-    }
-}
-
-impl FieldMapValueDocument {
-    fn into_document<E>(self, key: String) -> Result<FieldDocument, E>
-    where
-        E: de::Error,
-    {
-        match self {
-            Self::Type(ty) => Ok(FieldDocument {
-                name: key,
-                api_name: None,
-                ty,
-                items: None,
-                fields: Vec::new(),
-                nullable: false,
-                id: false,
-                generated: GeneratedValue::None,
-                unique: false,
-                transforms: Vec::new(),
-                relation: None,
-                validate: None,
-                legacy_validate: None,
-            }),
-            Self::Config(field) => field.into_document::<E>(key),
-        }
-    }
-}
-
-impl FieldMapConfigDocument {
-    fn into_document<E>(self, key: String) -> Result<FieldDocument, E>
-    where
-        E: de::Error,
-    {
-        if let Some(name) = self.name.as_deref()
-            && name != key
-        {
-            return Err(E::custom(format!(
-                "field map entry `{key}` has mismatched `name` value `{name}`"
-            )));
-        }
-
-        Ok(FieldDocument {
-            name: key,
-            api_name: self.api_name,
-            ty: self.ty,
-            items: self.items,
-            fields: self.fields,
-            nullable: self.nullable,
-            id: self.id,
-            generated: self.generated,
-            unique: self.unique,
-            transforms: self.transforms,
-            relation: self.relation,
-            validate: self.validate,
-            legacy_validate: self.legacy_validate,
-        })
-    }
-}
-
-impl MixinConfigDocument {
-    fn into_document<E>(self, key: String) -> Result<MixinDocument, E>
-    where
-        E: de::Error,
-    {
-        if let Some(name) = self.name.as_deref()
-            && name != key
-        {
-            return Err(E::custom(format!(
-                "mixin map entry `{key}` has mismatched `name` value `{name}`"
-            )));
-        }
-
-        Ok(MixinDocument {
-            name: key,
-            indexes: self.indexes,
-            fields: self.fields,
-        })
-    }
-}
-
-impl ApiFieldProjectionMapValueDocument {
-    fn into_document<E>(self, key: String) -> Result<ApiFieldProjectionDocument, E>
-    where
-        E: de::Error,
-    {
-        match self {
-            Self::From(from) => Ok(ApiFieldProjectionDocument {
-                name: key,
-                from: Some(from),
-                template: None,
-            }),
-            Self::Config(config) => config.into_document::<E>(key),
-        }
-    }
-}
-
-impl ApiFieldProjectionConfigDocument {
-    fn into_document<E>(self, key: String) -> Result<ApiFieldProjectionDocument, E>
-    where
-        E: de::Error,
-    {
-        if let Some(name) = self.name.as_deref()
-            && name != key
-        {
-            return Err(E::custom(format!(
-                "api.fields map entry `{key}` has mismatched `name` value `{name}`"
-            )));
-        }
-
-        Ok(ApiFieldProjectionDocument {
-            name: key,
-            from: self.from,
-            template: self.template,
-        })
-    }
-}
-
-impl ResponseContextMapValueDocument {
-    fn into_document<E>(self, key: String) -> Result<ResponseContextDocument, E>
-    where
-        E: de::Error,
-    {
-        match self {
-            Self::Fields(fields) => Ok(ResponseContextDocument { name: key, fields }),
-            Self::Config(config) => config.into_document::<E>(key),
-        }
-    }
-}
-
-impl ResponseContextConfigDocument {
-    fn into_document<E>(self, key: String) -> Result<ResponseContextDocument, E>
-    where
-        E: de::Error,
-    {
-        if let Some(name) = self.name.as_deref()
-            && name != key
-        {
-            return Err(E::custom(format!(
-                "api.contexts map entry `{key}` has mismatched `name` value `{name}`"
-            )));
-        }
-
-        Ok(ResponseContextDocument {
-            name: key,
-            fields: self.fields,
-        })
-    }
-}
-
-impl EnumMapValueDocument {
-    fn into_document<E>(self, key: String) -> Result<EnumDocument, E>
-    where
-        E: de::Error,
-    {
-        match self {
-            Self::Values(values) => Ok(EnumDocument { name: key, values }),
-            Self::Config(config) => config.into_document::<E>(key),
-        }
-    }
-}
-
-impl EnumConfigDocument {
-    fn into_document<E>(self, key: String) -> Result<EnumDocument, E>
-    where
-        E: de::Error,
-    {
-        if let Some(name) = self.name.as_deref()
-            && name != key
-        {
-            return Err(E::custom(format!(
-                "enums map entry `{key}` has mismatched `name` value `{name}`"
-            )));
-        }
-
-        Ok(EnumDocument {
-            name: key,
-            values: self.values,
-        })
-    }
-}
-
-struct ResourceDocumentsVisitor;
-
-struct EnumDocumentsVisitor;
-
-struct MixinDocumentsVisitor;
-
-impl<'de> Visitor<'de> for EnumDocumentsVisitor {
-    type Value = Vec<EnumDocument>;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.write_str("a list or map of enum definitions")
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: serde::de::SeqAccess<'de>,
-    {
-        let mut enums = Vec::new();
-        while let Some(document) = seq.next_element::<EnumDocument>()? {
-            enums.push(document);
-        }
-        Ok(enums)
-    }
-
-    fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-    where
-        A: MapAccess<'de>,
-    {
-        let mut seen = HashSet::new();
-        let mut enums = Vec::new();
-
-        while let Some((key, value)) = map.next_entry::<String, EnumMapValueDocument>()? {
-            if !seen.insert(key.clone()) {
-                return Err(A::Error::custom(format!("duplicate enum `{key}`")));
-            }
-            enums.push(value.into_document::<A::Error>(key)?);
-        }
-
-        Ok(enums)
-    }
-}
-
-impl<'de> Visitor<'de> for MixinDocumentsVisitor {
-    type Value = Vec<MixinDocument>;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.write_str("a list or map of mixin definitions")
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: serde::de::SeqAccess<'de>,
-    {
-        let mut mixins = Vec::new();
-        while let Some(document) = seq.next_element::<MixinDocument>()? {
-            mixins.push(document);
-        }
-        Ok(mixins)
-    }
-
-    fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-    where
-        A: MapAccess<'de>,
-    {
-        let mut seen = HashSet::new();
-        let mut mixins = Vec::new();
-
-        while let Some((key, value)) = map.next_entry::<String, MixinMapValueDocument>()? {
-            if !seen.insert(key.clone()) {
-                return Err(A::Error::custom(format!("duplicate mixin `{key}`")));
-            }
-            mixins.push(value.into_document::<A::Error>(key)?);
-        }
-
-        Ok(mixins)
-    }
-}
-
-impl<'de> Visitor<'de> for ResourceDocumentsVisitor {
-    type Value = Vec<ResourceDocument>;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.write_str("a list or map of resource definitions")
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: serde::de::SeqAccess<'de>,
-    {
-        let mut resources = Vec::new();
-        while let Some(resource) = seq.next_element::<ResourceDocument>()? {
-            resources.push(resource);
-        }
-        Ok(resources)
-    }
-
-    fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-    where
-        A: MapAccess<'de>,
-    {
-        let mut seen = HashSet::new();
-        let mut resources = Vec::new();
-
-        while let Some((key, value)) = map.next_entry::<String, ResourceMapValueDocument>()? {
-            if !seen.insert(key.clone()) {
-                return Err(A::Error::custom(format!("duplicate resource `{key}`")));
-            }
-            resources.push(value.into_document::<A::Error>(key)?);
-        }
-
-        Ok(resources)
-    }
-}
-
-struct FieldDocumentsVisitor;
-
-impl<'de> Visitor<'de> for FieldDocumentsVisitor {
-    type Value = Vec<FieldDocument>;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.write_str("a list or map of field definitions")
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: serde::de::SeqAccess<'de>,
-    {
-        let mut fields = Vec::new();
-        while let Some(field) = seq.next_element::<FieldDocument>()? {
-            fields.push(field);
-        }
-        Ok(fields)
-    }
-
-    fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-    where
-        A: MapAccess<'de>,
-    {
-        let mut seen = HashSet::new();
-        let mut fields = Vec::new();
-
-        while let Some((key, value)) = map.next_entry::<String, FieldMapValueDocument>()? {
-            if !seen.insert(key.clone()) {
-                return Err(A::Error::custom(format!("duplicate field `{key}`")));
-            }
-            fields.push(value.into_document::<A::Error>(key)?);
-        }
-
-        Ok(fields)
-    }
-}
-
-struct ApiFieldProjectionDocumentsVisitor;
-
-impl<'de> Visitor<'de> for ApiFieldProjectionDocumentsVisitor {
-    type Value = Vec<ApiFieldProjectionDocument>;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.write_str("a list or map of api field projection definitions")
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: serde::de::SeqAccess<'de>,
-    {
-        let mut fields = Vec::new();
-        while let Some(field) = seq.next_element::<ApiFieldProjectionDocument>()? {
-            fields.push(field);
-        }
-        Ok(fields)
-    }
-
-    fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-    where
-        A: MapAccess<'de>,
-    {
-        let mut seen = HashSet::new();
-        let mut fields = Vec::new();
-
-        while let Some((key, value)) =
-            map.next_entry::<String, ApiFieldProjectionMapValueDocument>()?
-        {
-            if !seen.insert(key.clone()) {
-                return Err(A::Error::custom(format!(
-                    "duplicate api field projection `{key}`"
-                )));
-            }
-            fields.push(value.into_document::<A::Error>(key)?);
-        }
-
-        Ok(fields)
-    }
-}
-
-struct ResponseContextDocumentsVisitor;
-
-impl<'de> Visitor<'de> for ResponseContextDocumentsVisitor {
-    type Value = Vec<ResponseContextDocument>;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        formatter.write_str("a list or map of response context definitions")
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: serde::de::SeqAccess<'de>,
-    {
-        let mut contexts = Vec::new();
-        while let Some(context) = seq.next_element::<ResponseContextDocument>()? {
-            contexts.push(context);
-        }
-        Ok(contexts)
-    }
-
-    fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-    where
-        A: MapAccess<'de>,
-    {
-        let mut seen = HashSet::new();
-        let mut contexts = Vec::new();
-
-        while let Some((key, value)) =
-            map.next_entry::<String, ResponseContextMapValueDocument>()?
-        {
-            if !seen.insert(key.clone()) {
-                return Err(A::Error::custom(format!(
-                    "duplicate response context `{key}`"
-                )));
-            }
-            contexts.push(value.into_document::<A::Error>(key)?);
-        }
-
-        Ok(contexts)
-    }
-}
+/// The result of loading and parsing a `.eon` service file.
+///
+/// Returned by [`load_service_from_file`] and [`load_service_from_path`].
+pub(super) struct LoadedService {
+    /// The parsed service specification.
+    pub(super) service: ServiceSpec,
+    /// Absolute path to the directory containing the `.eon` file, used to
+    /// resolve relative include paths during code generation.
+    pub(super) include_path: String,
+}
+
+mod documents;
+use self::documents::*;
 
 pub fn load_service_from_file(path: LitStr) -> syn::Result<LoadedService> {
     let manifest_dir = env::var("CARGO_MANIFEST_DIR")
